Sequence table try - failing

+++ src/main/java/no/kristiania/database/Task.java
@@ -2,16 +2,16 @@ package no.kristiania.database;
 
 public class Task {
     private String name;
-    private Integer id;
+    private Long id;
     private String ColorCode;
 
     public String getName() { return name; }
 
     public void setName(String name) { this.name = name; }
 
-    public void setId(Integer id) { this.id = id; }
+    public void setId(Long id) { this.id = id; }
 
-    public Integer getId() { return id; }
+    public Long getId() { return id; }
 
     public String getColorCode() { return ColorCode; }
