+++ src/main/java/no/kristiania/database/Worker.java
@@ -4,8 +4,8 @@ public class Worker {
     private String firstName;
     private String lastName;
     private String emailAddress;
-    private Integer id;
-    private Integer taskId;
+    private Long id;
+    private Long taskId;
 
     public String getFirstName() {
         return firstName;
@@ -15,11 +15,11 @@ public class Worker {
         this.firstName = firstName;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -39,11 +39,11 @@ public class Worker {
         this.emailAddress = emailAddress;
     }
 
-    public Integer getTaskId() {
+    public Long getTaskId() {
         return taskId;
     }
 
-    public void setTaskId(Integer taskId) {
+    public void setTaskId(Long taskId) {
         this.taskId = taskId;
     }
 }
